requirements.txt
# Azure AI Foundry Cost Pipeline - Dependencies
# ═══════════════════════════════════════════════════════════════════════════

# HTTP requests
requests>=2.28.0

# Environment variable management
python-dotenv>=1.0.0

# Data processing (for CSV mode)
pandas>=2.0.0

# ═══════════════════════════════════════════════════════════════════════════
# Note: The following are available in Databricks Runtime (DBR)
# and should NOT be installed locally (only needed for Delta mode):
#   - pyspark
#   - delta-spark
# ═══════════════════════════════════════════════════════════════════════════
